Doggo - project finished (without Flyway)

+++ doggo/src/main/java/doggo/dog/DogService.java
@@ -2,12 +2,13 @@ package doggo.dog;
 
 import doggo.NotFoundException;
 import doggo.owner.CreateOwnerCommand;
-import doggo.owner.OwnerDTO;
-import doggo.owner.UpdateWithExistingDogCommand;
+import doggo.owner.Owner;
+import doggo.owner.OwnerRepository;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.stereotype.Service;
+import javax.transaction.Transactional;
 import java.lang.reflect.Type;
 import java.util.List;
 import java.util.Optional;
@@ -20,34 +21,55 @@ public class DogService {
 
     private ModelMapper modelMapper;
 
+    private OwnerRepository ownerRepository;
+
 
     public List<DogDTO> listDogs(Optional<String> breed) {
         Type targetListType = new TypeToken<List<DogDTO>>(){}.getType();
-        return modelMapper.map(dogRepository.findAll(), targetListType);
-    }
+        return modelMapper.map(dogRepository.findAll(), targetListType);}
+
 
     public DogDTO findDogById(int id) {
-        Dog dog = dogRepository.findById(id).orElseThrow(()-> new NotFoundException(id));
-        return modelMapper.map(dog, DogDTO.class);
-    }
+        return modelMapper.map(findDog(id), DogDTO.class);}
 
-//    public List<DogDTO> listDogsByBreed(String breed) {
-//        Type targetListType = new TypeToken<List<DogDTO>>(){}.getType();
-//        List<Dog> result = dogRepository.findAll().stream().filter(d -> d.getBreed().equals(breed)).collect(Collectors.toList());
-//        return modelMapper.map(result, targetListType);
-//    }
+
+    public Dog findDog(int id) {
+        return dogRepository.findById(id).orElseThrow(()-> new NotFoundException(id, "Dog"));}
+
 
     public DogDTO saveDog(CreateDogCommand command) {
-        Dog dog = new Dog(command.getName(), command.getBreed(), command.getAge());
+        Dog dog = new Dog(command.getName(), command.getBreed(), command.getAge(), command.getFavToy());
         dogRepository.save(dog);
-        return modelMapper.map(dog, DogDTO.class);
-    }
+        return modelMapper.map(dog, DogDTO.class);}
+
+
+    @Transactional
+    public DogDTO addOwner(int id, CreateOwnerCommand command) {
+        Owner owner = new Owner(command.getName());
+        Dog dog = dogRepository.findById(id).orElseThrow(()-> new NotFoundException(id, "Dog"));
+        ownerRepository.save(owner);
+        dog.setOwner(owner);
+        return modelMapper.map(dog, DogDTO.class);}
+
+
+    @Transactional
+    public DogDTO updateDog(int id, UpdateDogCommand command) {
+        Dog dog = findDog(id);
+        dog.setName(command.getName());
+        dog.setAge(command.getAge());
+        dog.setBreed(command.getBreed());
+        dog.setFavToy(command.getFavToy());
+        Dog result = dogRepository.save(dog);
+        return modelMapper.map(result, DogDTO.class);}
+
 
     public void deleteDogById(int id) {
-        dogRepository.deleteById(id);
-    }
+        if(dogRepository.existsById(id)) {
+            dogRepository.deleteById(id);
+        } else {
+            throw new NotFoundException(id, "Dog");}}
+
 
     public void deleteAllTheDogs() {
         dogRepository.deleteAll();
-    }
-}
+    }}
